Added check for dimensions of image input
